requirements.txt
streamlit>=1.34.0,<2.0.0
PyMuPDF==1.23.26
python-docx==1.1.0
transformers==4.40.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
langchain==0.1.17
langchain-community==0.0.37
huggingface-hub==0.23.0
pillow
